Updated for betanet

+++ Sources/EngineToolkit/Models/Transaction/TransactionHeader.swift
@@ -16,7 +16,7 @@ public struct TransactionHeader: Sendable, Codable, Hashable {
     public let publicKey: Engine.PublicKey
     public let notaryAsSignatory: Bool
     public let costUnitLimit: UInt32
-    public let tipPercentage: UInt32
+    public let tipPercentage: UInt8
     
     private enum CodingKeys: String, CodingKey {
         case version = "version"
@@ -39,7 +39,7 @@ public struct TransactionHeader: Sendable, Codable, Hashable {
         publicKey: Engine.PublicKey,
         notaryAsSignatory: Bool,
         costUnitLimit: UInt32,
-        tipPercentage: UInt32
+        tipPercentage: UInt8
     ) {
         self.version = version
         self.networkId = networkId
